學員作品/Web_交通事故分析_04-毛栩櫟、16-陳名畯、19-林佑穎/requirements.txt
psycopg2-binary
python-dotenv
flask
gunicorn
dash
pandas
plotly.express
sqlalchemy